+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "axios": "^1.4.0",
     "node-sass-import": "^2.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
